implemented ownership of courses

+++ coursequery-backend/controllers/courses.js
@@ -32,6 +32,9 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @route   POST /api/v1/courses
 // @access  Private
 exports.createCourse = asyncHandler(async (req, res, next) => {
+  // Add user to req.body 
+  req.body.user = req.user.id;
+
   const course = await Course.create(req.body);
   res.status(201).json({
     success: true,
@@ -43,16 +46,26 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
 // @route   POST /api/v1/courses/:id
 // @access  Private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
-  const course = await Course.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidators: true,
-  });
+  let course = await Course.findById(req.params.id); 
 
   if (!course) {
     return next(
       new ErrorResponse(`Course not found with the id of ${req.params.id}`, 400)
     );
   }
+
+  // Check if user is the creator of the course
+  if(course.user.toString() !== req.user.id && req.user.role !== 'admin'){
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorized to update this course`, 401)
+    ); 
+  }
+
+  course = await Course.findByIdAndUpdate(req.params.id, req.body, {
+    new: true, 
+    runValidators: true
+  });
+
 
   res.status(200).json({ success: true, data: course });
 });
@@ -61,9 +74,18 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 // @route   DELETE /api/v1/courses/:id
 // @access  Private
 exports.deleteCourse = asyncHandler(async(req, res, next) => {
-    const course = await Course.findByIdAndDelete(req.params.id);
+    const course = await Course.findById(req.params.id);
     if(!course){
         return res.status(400).json({ success: false });
     }
+
+    if(course.user.toString() !== req.user.id && req.user.role !== 'admin'){
+      return next(
+        new ErrorResponse(`User ${req.params.id} is not authorized to delete this course`, 401)
+      ); 
+    }
+
+    await Course.findByIdAndDelete(req.params.id);
+
     res.status(200).json({ success: true, data: {} });
 })
